refactor: reducerのファイル名を変更
